+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Internal Changes
 - Skip ``twine check`` test on macOS until shell kill error is fixed. By `Kyle Brindley`_.
 - Restructure internal deployment around Gitlab-CI tag pipelines. By `Kyle Brindley`_.
 - Preserve QOI datatree structure when merging (:merge:`1205`). By `Matthew Fister`_.
+- Allow a QOI date attribute to be specified via the ``archive`` subcommand (:merge:`1211`). By `Matthew Fister`_.
 
 *******************
 0.13.4 (2025-05-29)